Overide config values from environment variables
The *Config.SyncEnv() overides the values of any field  that is set in the
environment. The method will look for names in the environment that
matches an uppercassed, underscore separated string of the corresponding
field name.

Example
  For the field AppName, the method will look for APP_NAME in the
environment.

Only int, string and bool fields are overidden, unsupported fields are
ignored.